src/requirements.txt
Appium-Python-Client==4.2.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
selenium==4.25.0
setuptools==75.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
